Create MULTIPLE DISEASE PREDICTION USING MACHINE LEARNING  APPROACH
In this project we analyze Diabetes analysis, heart disease and Parkinson’s disease by using 
some of the basic parameters such as Pulse Rate, Cholesterol, Blood Pressure, Heart Rate, etc., 
and also the risk factors associated with the disease can be found using prediction model with 
good accuracy and Precision. 
To implement multiple disease analysis using machine learning algorithms, Streamlit and 
python pickling is utilized to save the model behavior.  Further we can include other kinds of 
chronic diseases, skin diseases and many others. 
The significance of this analysis is to analyze the maximum diseases to screen the patient's 
condition and caution the patients ahead of time to diminish mortality proportion. 
In this task we are utilizing Naïve Bayes algorithm, random forest, decision tree and SVM 
classifier are utilized for prediction of a particular disease.  
The user has to enter various parameters of the disease and the system would display the output 
whether he/she has the disease or not. This project can help a lot of people as one can monitor 
the persons’ condition and take the necessary precautions thus increasing the life expectancy. 
The project emphasizes the interpretability and explain ability of the models, facilitating 
seamless integration into clinical decision-making processes. Extensive validation and 
performance assessment are conducted using diverse datasets to ensure the generalizability and 
reliability of the predictive models across various patient populations.